requirements.txt
requests
openai
python-dotenv
schedule
datetime
pandas
snowflake-connector-python